Refactor Input atom
