+++ src/components/atoms/Input.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+import { StyledInput } from "../../assets/styles/atoms/InputStyling";
+
+const Input = ({ ...inputProps }) => {
+  return <StyledInput {...inputProps} />;
+};
+
+export default Input;
